switched from float64 to int64 in arguments

+++ drive_array.go
@@ -3,15 +3,15 @@ package metalcloud
 import "log"
 
 type DriveArray struct {
-	DriveArrayID float64 `json:"drive_array_id,omitempty"`
+	DriveArrayID int64 `json:"drive_array_id,omitempty"`
 	DriveArrayLabel  string  `json:"drive_array_label,omitempty"`
-	VolumeTemplateID  float64  `json:"volume_template_id,omitempty"`
+	VolumeTemplateID  int64  `json:"volume_template_id,omitempty"`
 	DriveArrayStorageType string  `json:"drive_array_storage_type,omitempty"`
-	DriveSizeMBytesDefault float64  `json:"drive_size_mbytes_default,omitempty"`
-	InstanceArrayID float64  `json:"instance_array_id,omitempty"`	
+	DriveSizeMBytesDefault int64  `json:"drive_size_mbytes_default,omitempty"`
+	InstanceArrayID int64  `json:"instance_array_id,omitempty"`	
 }
 
-func (c *MetalCloudClient) DriveArrays(instanceArrayID float64) (*map[string]DriveArray, error) {
+func (c *MetalCloudClient) DriveArrays(instanceArrayID int64) (*map[string]DriveArray, error) {
 	var created_object map[string]DriveArray
 
 	err := c.rpcClient.CallFor(
@@ -26,7 +26,7 @@ func (c *MetalCloudClient) DriveArrays(instanceArrayID float64) (*map[string]Dri
 	return &created_object, nil
 }
 
-func (c *MetalCloudClient) DriveArrayGet(driveArrayID float64) (*DriveArray, error) {
+func (c *MetalCloudClient) DriveArrayGet(driveArrayID int64) (*DriveArray, error) {
 	var created_object DriveArray
 
 	err := c.rpcClient.CallFor(
@@ -42,7 +42,7 @@ func (c *MetalCloudClient) DriveArrayGet(driveArrayID float64) (*DriveArray, err
 	return &created_object, nil
 }
 
-func (c *MetalCloudClient) DriveArrayCreate(infrastructureID float64, driveArray DriveArray) (*DriveArray, error) {
+func (c *MetalCloudClient) DriveArrayCreate(infrastructureID int64, driveArray DriveArray) (*DriveArray, error) {
 	var created_object DriveArray
 
 	err := c.rpcClient.CallFor(
